// client/src/model/house.ts
export class House {
  id: number = 0;
  levelCount: number = 0;
  ceiling: number = 0;
  roomCount: number = 0;
  landArea: number = 0;
  area: number = 0;
  water: string = "";
  gas: string = "";
  electricity: string = "";
  sewer: string = "";
  walls: string = "";
  hotWater: string = "";
  condition: string = "";
  heating: string = "";
  bath: string = "";
  toilet: string = "";
  hasGarage: boolean = false;
  constructionTime: string = "";
  roof: string = "";
}