// client/src/model/options/house-options.ts
export class HouseOptions {
  water: string[] = [];
  gas: string[] = [];
  electricity: string[] = [];
  sewer: string[] = [];
  walls: string[] = [];
  hotWater: string[] = [];
  condition: string[] = [];
  heating: string[] = [];
  bath: string[] = [];
  toilet: string[] = [];
  constructionTime: string[] = [];
  roof: string[] = [];
}